Phase 4 completed
