+++ ICSharpCode.AvalonEdit/Editing/TextArea.cs
@@ -1,14 +1,14 @@
-﻿// Copyright (c) 2014 AlphaSierraPapa for the SharpDevelop Team
-// 
+// Copyright (c) 2014 AlphaSierraPapa for the SharpDevelop Team
+//
 // Permission is hereby granted, free of charge, to any person obtaining a copy of this
 // software and associated documentation files (the "Software"), to deal in the Software
 // without restriction, including without limitation the rights to use, copy, modify, merge,
 // publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons
 // to whom the Software is furnished to do so, subject to the following conditions:
-// 
+//
 // The above copyright notice and this permission notice shall be included in all copies or
 // substantial portions of the Software.
-// 
+//
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
 // INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
 // PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE
@@ -24,10 +24,8 @@ using System.Diagnostics;
 using System.Linq;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Controls.Primitives;
 using System.Windows.Documents;
 using System.Windows.Input;
-using System.Windows.Media;
 using System.Windows.Threading;
 
 using ICSharpCode.AvalonEdit.Document;
@@ -40,21 +38,24 @@ namespace ICSharpCode.AvalonEdit.Editing
 	/// <summary>
 	/// Control that wraps a TextView and adds support for user input and the caret.
 	/// </summary>
-	public class TextArea : Control, IScrollInfo, IWeakEventListener, ITextEditorComponent, IServiceProvider
+	public partial class TextArea : Control, ITextEditorComponent, IServiceProvider
 	{
-		internal readonly ImeSupport ime;
+		#region Partial void declarations
+		static partial void InitializeWpfDefaults();
+		partial void InitializeIme();
+		partial void AttachTypingEvents();
+		partial void RequestSelectionValidationAsync();
+		partial void InvalidateRequery();
+		partial void SubscribeToDocumentEvents(TextDocument document);
+		partial void UnsubscribeFromDocumentEvents(TextDocument document);
+		partial void SubscribeToOptionsEvents(TextEditorOptions options);
+		partial void UnsubscribeFromOptionsEvents(TextEditorOptions options);
+		#endregion
 
 		#region Constructor
 		static TextArea()
 		{
-			DefaultStyleKeyProperty.OverrideMetadata(typeof(TextArea),
-													 new FrameworkPropertyMetadata(typeof(TextArea)));
-			KeyboardNavigation.IsTabStopProperty.OverrideMetadata(
-				typeof(TextArea), new FrameworkPropertyMetadata(Boxes.True));
-			KeyboardNavigation.TabNavigationProperty.OverrideMetadata(
-				typeof(TextArea), new FrameworkPropertyMetadata(KeyboardNavigationMode.None));
-			FocusableProperty.OverrideMetadata(
-				typeof(TextArea), new FrameworkPropertyMetadata(Boxes.True));
+			InitializeWpfDefaults();
 		}
 
 		/// <summary>
@@ -85,7 +86,7 @@ namespace ICSharpCode.AvalonEdit.Editing
 			caret.PositionChanged += (sender, e) => RequestSelectionValidation();
 			caret.PositionChanged += CaretPositionChanged;
 			AttachTypingEvents();
-			ime = new ImeSupport(this);
+			InitializeIme();
 
 			leftMargins.CollectionChanged += leftMargins_CollectionChanged;
 
@@ -208,26 +209,16 @@ namespace ICSharpCode.AvalonEdit.Editing
 
 		void OnDocumentChanged(TextDocument oldValue, TextDocument newValue)
 		{
-			if (oldValue != null) {
-				TextDocumentWeakEventManager.Changing.RemoveListener(oldValue, this);
-				TextDocumentWeakEventManager.Changed.RemoveListener(oldValue, this);
-				TextDocumentWeakEventManager.UpdateStarted.RemoveListener(oldValue, this);
-				TextDocumentWeakEventManager.UpdateFinished.RemoveListener(oldValue, this);
-			}
+			UnsubscribeFromDocumentEvents(oldValue);
 			textView.Document = newValue;
-			if (newValue != null) {
-				TextDocumentWeakEventManager.Changing.AddListener(newValue, this);
-				TextDocumentWeakEventManager.Changed.AddListener(newValue, this);
-				TextDocumentWeakEventManager.UpdateStarted.AddListener(newValue, this);
-				TextDocumentWeakEventManager.UpdateFinished.AddListener(newValue, this);
-			}
+			SubscribeToDocumentEvents(newValue);
 			// Reset caret location and selection: this is necessary because the caret/selection might be invalid
 			// in the new document (e.g. if new document is shorter than the old document).
 			caret.Location = new TextLocation(1, 1);
 			this.ClearSelection();
 			if (DocumentChanged != null)
 				DocumentChanged(this, EventArgs.Empty);
-			CommandManager.InvalidateRequerySuggested();
+			InvalidateRequery();
 		}
 		#endregion
 
@@ -268,43 +259,10 @@ namespace ICSharpCode.AvalonEdit.Editing
 
 		void OnOptionsChanged(TextEditorOptions oldValue, TextEditorOptions newValue)
 		{
-			if (oldValue != null) {
-				PropertyChangedWeakEventManager.RemoveListener(oldValue, this);
-			}
+			UnsubscribeFromOptionsEvents(oldValue);
 			textView.Options = newValue;
-			if (newValue != null) {
-				PropertyChangedWeakEventManager.AddListener(newValue, this);
-			}
+			SubscribeToOptionsEvents(newValue);
 			OnOptionChanged(new PropertyChangedEventArgs(null));
-		}
-		#endregion
-
-		#region ReceiveWeakEvent
-		/// <inheritdoc cref="IWeakEventListener.ReceiveWeakEvent"/>
-		protected virtual bool ReceiveWeakEvent(Type managerType, object sender, EventArgs e)
-		{
-			if (managerType == typeof(TextDocumentWeakEventManager.Changing)) {
-				OnDocumentChanging();
-				return true;
-			} else if (managerType == typeof(TextDocumentWeakEventManager.Changed)) {
-				OnDocumentChanged((DocumentChangeEventArgs)e);
-				return true;
-			} else if (managerType == typeof(TextDocumentWeakEventManager.UpdateStarted)) {
-				OnUpdateStarted();
-				return true;
-			} else if (managerType == typeof(TextDocumentWeakEventManager.UpdateFinished)) {
-				OnUpdateFinished();
-				return true;
-			} else if (managerType == typeof(PropertyChangedWeakEventManager)) {
-				OnOptionChanged((PropertyChangedEventArgs)e);
-				return true;
-			}
-			return false;
-		}
-
-		bool IWeakEventListener.ReceiveWeakEvent(Type managerType, object sender, EventArgs e)
-		{
-			return ReceiveWeakEvent(managerType, sender, e);
 		}
 		#endregion
 
@@ -365,7 +323,6 @@ namespace ICSharpCode.AvalonEdit.Editing
 
 		#region TextView property
 		readonly TextView textView;
-		IScrollInfo scrollInfo;
 
 		/// <summary>
 		/// Gets the text view used to display text in this text area.
@@ -374,13 +331,6 @@ namespace ICSharpCode.AvalonEdit.Editing
 			get {
 				return textView;
 			}
-		}
-		/// <inheritdoc/>
-		public override void OnApplyTemplate()
-		{
-			base.OnApplyTemplate();
-			scrollInfo = textView;
-			ApplyScrollInfo();
 		}
 		#endregion
 
@@ -434,7 +384,7 @@ namespace ICSharpCode.AvalonEdit.Editing
 					if (SelectionChanged != null)
 						SelectionChanged(this, EventArgs.Empty);
 					// a selection change causes commands like copy/paste/etc. to change status
-					CommandManager.InvalidateRequerySuggested();
+					InvalidateRequery();
 				}
 			}
 		}
@@ -451,13 +401,13 @@ namespace ICSharpCode.AvalonEdit.Editing
 		/// The <see cref="SelectionBrush"/> property.
 		/// </summary>
 		public static readonly DependencyProperty SelectionBrushProperty =
-			DependencyProperty.Register("SelectionBrush", typeof(Brush), typeof(TextArea));
+			DependencyProperty.Register("SelectionBrush", typeof(System.Windows.Media.Brush), typeof(TextArea));
 
 		/// <summary>
 		/// Gets/Sets the background brush used for the selection.
 		/// </summary>
-		public Brush SelectionBrush {
-			get { return (Brush)GetValue(SelectionBrushProperty); }
+		public System.Windows.Media.Brush SelectionBrush {
+			get { return (System.Windows.Media.Brush)GetValue(SelectionBrushProperty); }
 			set { SetValue(SelectionBrushProperty, value); }
 		}
 
@@ -465,13 +415,13 @@ namespace ICSharpCode.AvalonEdit.Editing
 		/// The <see cref="SelectionForeground"/> property.
 		/// </summary>
 		public static readonly DependencyProperty SelectionForegroundProperty =
-			DependencyProperty.Register("SelectionForeground", typeof(Brush), typeof(TextArea));
+			DependencyProperty.Register("SelectionForeground", typeof(System.Windows.Media.Brush), typeof(TextArea));
 
 		/// <summary>
 		/// Gets/Sets the foreground brush used for selected text.
 		/// </summary>
-		public Brush SelectionForeground {
-			get { return (Brush)GetValue(SelectionForegroundProperty); }
+		public System.Windows.Media.Brush SelectionForeground {
+			get { return (System.Windows.Media.Brush)GetValue(SelectionForegroundProperty); }
 			set { SetValue(SelectionForegroundProperty, value); }
 		}
 
@@ -479,13 +429,13 @@ namespace ICSharpCode.AvalonEdit.Editing
 		/// The <see cref="SelectionBorder"/> property.
 		/// </summary>
 		public static readonly DependencyProperty SelectionBorderProperty =
-			DependencyProperty.Register("SelectionBorder", typeof(Pen), typeof(TextArea));
+			DependencyProperty.Register("SelectionBorder", typeof(System.Windows.Media.Pen), typeof(TextArea));
 
 		/// <summary>
 		/// Gets/Sets the pen used for the border of the selection.
 		/// </summary>
-		public Pen SelectionBorder {
-			get { return (Pen)GetValue(SelectionBorderProperty); }
+		public System.Windows.Media.Pen SelectionBorder {
+			get { return (System.Windows.Media.Pen)GetValue(SelectionBorderProperty); }
 			set { SetValue(SelectionBorderProperty, value); }
 		}
 
@@ -506,14 +456,14 @@ namespace ICSharpCode.AvalonEdit.Editing
 
 		/// <summary>
 		/// Gets/Sets the active mouse selection mode.
-		/// 
+		///
 		/// Setting this property to MouseSelectionMode.None will cancel mouse selection
 		/// and release mouse capture.
-		/// 
+		///
 		/// Setting this property to another value will acquire mouse capture and
 		/// activate the mouse selection mode.
 		/// If mouse capture cannot be acquired, MouseSelectionMode will stay unchanged.
-		/// 
+		///
 		/// Currently, the setter only supports the values <c>None</c>, <c>Normal</c>
 		/// and <c>Rectangular</c>.
 		/// </summary>
@@ -543,7 +493,7 @@ namespace ICSharpCode.AvalonEdit.Editing
 		{
 			if (!ensureSelectionValidRequested && allowCaretOutsideSelection == 0) {
 				ensureSelectionValidRequested = true;
-				Dispatcher.BeginInvoke(DispatcherPriority.Normal, new Action(EnsureSelectionValid));
+				RequestSelectionValidationAsync();
 			}
 		}
 
@@ -551,10 +501,10 @@ namespace ICSharpCode.AvalonEdit.Editing
 		/// Code that updates only the caret but not the selection can cause confusion when
 		/// keys like 'Delete' delete the (possibly invisible) selected text and not the
 		/// text around the caret.
-		/// 
+		///
 		/// So we'll ensure that the caret is inside the selection.
 		/// (when the caret is not in the selection, we'll clear the selection)
-		/// 
+		///
 		/// This method is invoked using the Dispatcher so that code may temporarily violate this rule
 		/// (e.g. most 'extend selection' methods work by first setting the caret, then the selection),
 		/// it's sufficient to fix it after any event handlers have run.
@@ -648,278 +598,12 @@ namespace ICSharpCode.AvalonEdit.Editing
 				if (value == null)
 					throw new ArgumentNullException("value");
 				readOnlySectionProvider = value;
-				CommandManager.InvalidateRequerySuggested(); // the read-only status effects Paste.CanExecute and the IME
+				InvalidateRequery(); // the read-only status effects Paste.CanExecute and the IME
 			}
-		}
-		#endregion
-
-		#region IScrollInfo implementation
-		ScrollViewer scrollOwner;
-		bool canVerticallyScroll, canHorizontallyScroll;
-
-		void ApplyScrollInfo()
-		{
-			if (scrollInfo != null) {
-				scrollInfo.ScrollOwner = scrollOwner;
-				scrollInfo.CanVerticallyScroll = canVerticallyScroll;
-				scrollInfo.CanHorizontallyScroll = canHorizontallyScroll;
-				scrollOwner = null;
-			}
-		}
-
-		bool IScrollInfo.CanVerticallyScroll {
-			get { return scrollInfo != null ? scrollInfo.CanVerticallyScroll : false; }
-			set {
-				canVerticallyScroll = value;
-				if (scrollInfo != null)
-					scrollInfo.CanVerticallyScroll = value;
-			}
-		}
-
-		bool IScrollInfo.CanHorizontallyScroll {
-			get { return scrollInfo != null ? scrollInfo.CanHorizontallyScroll : false; }
-			set {
-				canHorizontallyScroll = value;
-				if (scrollInfo != null)
-					scrollInfo.CanHorizontallyScroll = value;
-			}
-		}
-
-		double IScrollInfo.ExtentWidth {
-			get { return scrollInfo != null ? scrollInfo.ExtentWidth : 0; }
-		}
-
-		double IScrollInfo.ExtentHeight {
-			get { return scrollInfo != null ? scrollInfo.ExtentHeight : 0; }
-		}
-
-		double IScrollInfo.ViewportWidth {
-			get { return scrollInfo != null ? scrollInfo.ViewportWidth : 0; }
-		}
-
-		double IScrollInfo.ViewportHeight {
-			get { return scrollInfo != null ? scrollInfo.ViewportHeight : 0; }
-		}
-
-		double IScrollInfo.HorizontalOffset {
-			get { return scrollInfo != null ? scrollInfo.HorizontalOffset : 0; }
-		}
-
-		double IScrollInfo.VerticalOffset {
-			get { return scrollInfo != null ? scrollInfo.VerticalOffset : 0; }
-		}
-
-		ScrollViewer IScrollInfo.ScrollOwner {
-			get { return scrollInfo != null ? scrollInfo.ScrollOwner : null; }
-			set {
-				if (scrollInfo != null)
-					scrollInfo.ScrollOwner = value;
-				else
-					scrollOwner = value;
-			}
-		}
-
-		void IScrollInfo.LineUp()
-		{
-			if (scrollInfo != null) scrollInfo.LineUp();
-		}
-
-		void IScrollInfo.LineDown()
-		{
-			if (scrollInfo != null) scrollInfo.LineDown();
-		}
-
-		void IScrollInfo.LineLeft()
-		{
-			if (scrollInfo != null) scrollInfo.LineLeft();
-		}
-
-		void IScrollInfo.LineRight()
-		{
-			if (scrollInfo != null) scrollInfo.LineRight();
-		}
-
-		void IScrollInfo.PageUp()
-		{
-			if (scrollInfo != null) scrollInfo.PageUp();
-		}
-
-		void IScrollInfo.PageDown()
-		{
-			if (scrollInfo != null) scrollInfo.PageDown();
-		}
-
-		void IScrollInfo.PageLeft()
-		{
-			if (scrollInfo != null) scrollInfo.PageLeft();
-		}
-
-		void IScrollInfo.PageRight()
-		{
-			if (scrollInfo != null) scrollInfo.PageRight();
-		}
-
-		void IScrollInfo.MouseWheelUp()
-		{
-			if (scrollInfo != null) scrollInfo.MouseWheelUp();
-		}
-
-		void IScrollInfo.MouseWheelDown()
-		{
-			if (scrollInfo != null) scrollInfo.MouseWheelDown();
-		}
-
-		void IScrollInfo.MouseWheelLeft()
-		{
-			if (scrollInfo != null) scrollInfo.MouseWheelLeft();
-		}
-
-		void IScrollInfo.MouseWheelRight()
-		{
-			if (scrollInfo != null) scrollInfo.MouseWheelRight();
-		}
-
-		void IScrollInfo.SetHorizontalOffset(double offset)
-		{
-			if (scrollInfo != null) scrollInfo.SetHorizontalOffset(offset);
-		}
-
-		void IScrollInfo.SetVerticalOffset(double offset)
-		{
-			if (scrollInfo != null) scrollInfo.SetVerticalOffset(offset);
-		}
-
-		Rect IScrollInfo.MakeVisible(System.Windows.Media.Visual visual, Rect rectangle)
-		{
-			if (scrollInfo != null)
-				return scrollInfo.MakeVisible(visual, rectangle);
-			else
-				return Rect.Empty;
-		}
-		#endregion
-
-		#region Focus Handling (Show/Hide Caret)
-		/// <inheritdoc/>
-		protected override void OnMouseDown(MouseButtonEventArgs e)
-		{
-			base.OnMouseDown(e);
-			Focus();
-		}
-
-		/// <inheritdoc/>
-		protected override void OnGotKeyboardFocus(KeyboardFocusChangedEventArgs e)
-		{
-			base.OnGotKeyboardFocus(e);
-			// First activate IME, then show caret
-			ime.OnGotKeyboardFocus(e);
-			caret.Show();
-		}
-
-		/// <inheritdoc/>
-		protected override void OnLostKeyboardFocus(KeyboardFocusChangedEventArgs e)
-		{
-			base.OnLostKeyboardFocus(e);
-			caret.Hide();
-			ime.OnLostKeyboardFocus(e);
 		}
 		#endregion
 
 		#region OnTextInput / RemoveSelectedText / ReplaceSelectionWithText
-		/// <summary>
-		/// Occurs when the TextArea receives text input.
-		/// This is like the <see cref="UIElement.TextInput"/> event,
-		/// but occurs immediately before the TextArea handles the TextInput event.
-		/// </summary>
-		public event TextCompositionEventHandler TextEntering;
-
-		/// <summary>
-		/// Occurs when the TextArea receives text input.
-		/// This is like the <see cref="UIElement.TextInput"/> event,
-		/// but occurs immediately after the TextArea handles the TextInput event.
-		/// </summary>
-		public event TextCompositionEventHandler TextEntered;
-
-		/// <summary>
-		/// Raises the TextEntering event.
-		/// </summary>
-		protected virtual void OnTextEntering(TextCompositionEventArgs e)
-		{
-			if (TextEntering != null) {
-				TextEntering(this, e);
-			}
-		}
-
-		/// <summary>
-		/// Raises the TextEntered event.
-		/// </summary>
-		protected virtual void OnTextEntered(TextCompositionEventArgs e)
-		{
-			if (TextEntered != null) {
-				TextEntered(this, e);
-			}
-		}
-
-		/// <inheritdoc/>
-		protected override void OnTextInput(TextCompositionEventArgs e)
-		{
-			//Debug.WriteLine("TextInput: Text='" + e.Text + "' SystemText='" + e.SystemText + "' ControlText='" + e.ControlText + "'");
-			base.OnTextInput(e);
-			if (!e.Handled && this.Document != null) {
-				if (string.IsNullOrEmpty(e.Text) || e.Text == "\x1b" || e.Text == "\b") {
-					// ASCII 0x1b = ESC.
-					// WPF produces a TextInput event with that old ASCII control char
-					// when Escape is pressed. We'll just ignore it.
-
-					// A deadkey followed by backspace causes a textinput event for the BS character.
-
-					// Similarly, some shortcuts like Alt+Space produce an empty TextInput event.
-					// We have to ignore those (not handle them) to keep the shortcut working.
-					return;
-				}
-				HideMouseCursor();
-				PerformTextInput(e);
-				e.Handled = true;
-			}
-		}
-
-		/// <summary>
-		/// Performs text input.
-		/// This raises the <see cref="TextEntering"/> event, replaces the selection with the text,
-		/// and then raises the <see cref="TextEntered"/> event.
-		/// </summary>
-		public void PerformTextInput(string text)
-		{
-			TextComposition textComposition = new TextComposition(InputManager.Current, this, text);
-			TextCompositionEventArgs e = new TextCompositionEventArgs(Keyboard.PrimaryDevice, textComposition);
-			e.RoutedEvent = TextInputEvent;
-			PerformTextInput(e);
-		}
-
-		/// <summary>
-		/// Performs text input.
-		/// This raises the <see cref="TextEntering"/> event, replaces the selection with the text,
-		/// and then raises the <see cref="TextEntered"/> event.
-		/// </summary>
-		public void PerformTextInput(TextCompositionEventArgs e)
-		{
-			if (e == null)
-				throw new ArgumentNullException("e");
-			if (this.Document == null)
-				throw ThrowUtil.NoDocumentAssigned();
-			OnTextEntering(e);
-			if (!e.Handled) {
-				if (e.Text == "\n" || e.Text == "\r" || e.Text == "\r\n")
-					ReplaceSelectionWithNewLine();
-				else {
-					if (OverstrikeMode && Selection.IsEmpty && Document.GetLineByNumber(Caret.Line).EndOffset > Caret.Offset)
-						EditingCommands.SelectRightByCharacter.Execute(null, this);
-					ReplaceSelectionWithText(e.Text);
-				}
-				OnTextEntered(e);
-				caret.BringCaretToView();
-			}
-		}
-
 		void ReplaceSelectionWithNewLine()
 		{
 			string newLine = TextUtilities.GetNewLineFromDocument(this.Document, this.Caret.Line);
@@ -975,6 +659,18 @@ namespace ICSharpCode.AvalonEdit.Editing
 				throw new InvalidOperationException("ReadOnlySectionProvider returned incorrect segments (outside of input segment / wrong order)");
 			return array;
 		}
+
+		/// <summary>Core text insertion used by both the shared path and the WPF TextComposition path.</summary>
+		internal void InsertText(string text)
+		{
+			if (text == "\n" || text == "\r" || text == "\r\n")
+				ReplaceSelectionWithNewLine();
+			else {
+				if (OverstrikeMode && Selection.IsEmpty && Document.GetLineByNumber(Caret.Line).EndOffset > Caret.Offset)
+					EditingCommands.SelectRightByCharacter.Execute(null, this);
+				ReplaceSelectionWithText(text);
+			}
+		}
 		#endregion
 
 		#region IndentationStrategy property
@@ -992,78 +688,6 @@ namespace ICSharpCode.AvalonEdit.Editing
 			get { return (IIndentationStrategy)GetValue(IndentationStrategyProperty); }
 			set { SetValue(IndentationStrategyProperty, value); }
 		}
-		#endregion
-
-		#region OnKeyDown/OnKeyUp
-		/// <inheritdoc/>
-		protected override void OnPreviewKeyDown(KeyEventArgs e)
-		{
-			base.OnPreviewKeyDown(e);
-			foreach (TextAreaStackedInputHandler h in stackedInputHandlers) {
-				if (e.Handled)
-					break;
-				h.OnPreviewKeyDown(e);
-			}
-		}
-
-		/// <inheritdoc/>
-		protected override void OnPreviewKeyUp(KeyEventArgs e)
-		{
-			base.OnPreviewKeyUp(e);
-			foreach (TextAreaStackedInputHandler h in stackedInputHandlers) {
-				if (e.Handled)
-					break;
-				h.OnPreviewKeyUp(e);
-			}
-		}
-
-		// Make life easier for text editor extensions that use a different cursor based on the pressed modifier keys.
-		/// <inheritdoc/>
-		protected override void OnKeyDown(KeyEventArgs e)
-		{
-			base.OnKeyDown(e);
-			TextView.InvalidateCursorIfMouseWithinTextView();
-		}
-
-		/// <inheritdoc/>
-		protected override void OnKeyUp(KeyEventArgs e)
-		{
-			base.OnKeyUp(e);
-			TextView.InvalidateCursorIfMouseWithinTextView();
-		}
-		#endregion
-
-		#region Hide Mouse Cursor While Typing
-
-		bool isMouseCursorHidden;
-
-		void AttachTypingEvents()
-		{
-			// Use the PreviewMouseMove event in case some other editor layer consumes the MouseMove event (e.g. SD's InsertionCursorLayer)
-			this.MouseEnter += delegate { ShowMouseCursor(); };
-			this.MouseLeave += delegate { ShowMouseCursor(); };
-			this.PreviewMouseMove += delegate { ShowMouseCursor(); };
-			this.TouchEnter += delegate { ShowMouseCursor(); };
-			this.TouchLeave += delegate { ShowMouseCursor(); };
-			this.PreviewTouchMove += delegate { ShowMouseCursor(); };
-		}
-
-		void ShowMouseCursor()
-		{
-			if (this.isMouseCursorHidden) {
-				System.Windows.Forms.Cursor.Show();
-				this.isMouseCursorHidden = false;
-			}
-		}
-
-		void HideMouseCursor()
-		{
-			if (Options.HideCursorWhileTyping && !this.isMouseCursorHidden && this.IsMouseOver) {
-				this.isMouseCursorHidden = true;
-				System.Windows.Forms.Cursor.Hide();
-			}
-		}
-
 		#endregion
 
 		#region Overstrike mode
@@ -1084,33 +708,6 @@ namespace ICSharpCode.AvalonEdit.Editing
 		}
 
 		#endregion
-
-		/// <inheritdoc/>
-		protected override System.Windows.Automation.Peers.AutomationPeer OnCreateAutomationPeer()
-		{
-			return new TextAreaAutomationPeer(this);
-		}
-
-		/// <inheritdoc/>
-		protected override HitTestResult HitTestCore(PointHitTestParameters hitTestParameters)
-		{
-			// accept clicks even where the text area draws no background
-			return new PointHitTestResult(this, hitTestParameters.HitPoint);
-		}
-
-		/// <inheritdoc/>
-		protected override void OnPropertyChanged(DependencyPropertyChangedEventArgs e)
-		{
-			base.OnPropertyChanged(e);
-			if (e.Property == SelectionBrushProperty
-				|| e.Property == SelectionBorderProperty
-				|| e.Property == SelectionForegroundProperty
-				|| e.Property == SelectionCornerRadiusProperty) {
-				textView.Redraw();
-			} else if (e.Property == OverstrikeModeProperty) {
-				caret.UpdateIfVisible();
-			}
-		}
 
 		/// <summary>
 		/// Gets the requested service.
