+++ ICSharpCode.AvalonEdit/TextEditor.cs
@@ -1,14 +1,14 @@
-﻿// Copyright (c) 2014 AlphaSierraPapa for the SharpDevelop Team
-// 
+// Copyright (c) 2014 AlphaSierraPapa for the SharpDevelop Team
+//
 // Permission is hereby granted, free of charge, to any person obtaining a copy of this
 // software and associated documentation files (the "Software"), to deal in the Software
 // without restriction, including without limitation the rights to use, copy, modify, merge,
 // publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons
 // to whom the Software is furnished to do so, subject to the following conditions:
-// 
+//
 // The above copyright notice and this permission notice shall be included in all copies or
 // substantial portions of the Software.
-// 
+//
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
 // INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
 // PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE
@@ -22,13 +22,7 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Controls.Primitives;
-using System.Windows.Data;
-using System.Windows.Input;
-using System.Windows.Markup;
 using System.Windows.Media;
-using System.Windows.Shapes;
 
 using ICSharpCode.AvalonEdit.Document;
 using ICSharpCode.AvalonEdit.Editing;
@@ -42,17 +36,22 @@ namespace ICSharpCode.AvalonEdit
 	/// The text editor control.
 	/// Contains a scrollable TextArea.
 	/// </summary>
-	[Localizability(LocalizationCategory.Text), ContentProperty("Text")]
-	public class TextEditor : Control, ITextEditorComponent, IServiceProvider, IWeakEventListener
+	public partial class TextEditor : ITextEditorComponent, IServiceProvider
 	{
 		#region Constructors
 		static TextEditor()
 		{
-			DefaultStyleKeyProperty.OverrideMetadata(typeof(TextEditor),
-													 new FrameworkPropertyMetadata(typeof(TextEditor)));
-			FocusableProperty.OverrideMetadata(typeof(TextEditor),
-											   new FrameworkPropertyMetadata(Boxes.True));
+			InitializeWpfDefaults();
 		}
+
+		static partial void InitializeWpfDefaults();
+
+		// Platform-specific hooks for property-changed callbacks.
+		partial void OnDocumentChangedCore(TextDocument oldValue, TextDocument newValue);
+		partial void OnOptionsChangedCore(TextEditorOptions oldValue, TextEditorOptions newValue);
+		partial void OnIsReadOnlyChangedCore(TextEditor editor, bool oldValue, bool newValue);
+		partial void OnShowLineNumbersChangedCore(TextEditor editor, bool oldValue, bool newValue);
+		partial void OnLineNumbersForegroundChangedCore(TextEditor editor, object newValue);
 
 		/// <summary>
 		/// Creates a new TextEditor instance.
@@ -75,33 +74,11 @@ namespace ICSharpCode.AvalonEdit
 			SetCurrentValue(OptionsProperty, textArea.Options);
 			SetCurrentValue(DocumentProperty, new TextDocument());
 		}
-
 		#endregion
 
-		/// <inheritdoc/>
-		protected override System.Windows.Automation.Peers.AutomationPeer OnCreateAutomationPeer()
-		{
-			return new TextEditorAutomationPeer(this);
-		}
-
-		/// Forward focus to TextArea.
-		/// <inheritdoc/>
-		protected override void OnGotKeyboardFocus(KeyboardFocusChangedEventArgs e)
-		{
-			base.OnGotKeyboardFocus(e);
-			if (e.NewFocus == this) {
-				Keyboard.Focus(textArea);
-				e.Handled = true;
-			}
-		}
-
 		#region Document property
-		/// <summary>
-		/// Document property.
-		/// </summary>
-		public static readonly DependencyProperty DocumentProperty
-			= TextView.DocumentProperty.AddOwner(
-				typeof(TextEditor), new FrameworkPropertyMetadata(OnDocumentChanged));
+		// DocumentProperty is declared in TextEditor.wpf.cs (WPF: AddOwner from TextView)
+		// or TextEditor.uno.cs (Uno: DependencyProperty.Register) for each platform.
 
 		/// <summary>
 		/// Gets/Sets the document displayed by the text editor.
@@ -134,26 +111,16 @@ namespace ICSharpCode.AvalonEdit
 
 		void OnDocumentChanged(TextDocument oldValue, TextDocument newValue)
 		{
-			if (oldValue != null) {
-				TextDocumentWeakEventManager.TextChanged.RemoveListener(oldValue, this);
-				PropertyChangedEventManager.RemoveListener(oldValue.UndoStack, this, "IsOriginalFile");
-			}
+			OnDocumentChangedCore(oldValue, newValue);
 			textArea.Document = newValue;
-			if (newValue != null) {
-				TextDocumentWeakEventManager.TextChanged.AddListener(newValue, this);
-				PropertyChangedEventManager.AddListener(newValue.UndoStack, this, "IsOriginalFile");
-			}
 			OnDocumentChanged(EventArgs.Empty);
 			OnTextChanged(EventArgs.Empty);
 		}
 		#endregion
 
 		#region Options property
-		/// <summary>
-		/// Options property.
-		/// </summary>
-		public static readonly DependencyProperty OptionsProperty
-			= TextView.OptionsProperty.AddOwner(typeof(TextEditor), new FrameworkPropertyMetadata(OnOptionsChanged));
+		// OptionsProperty is declared in TextEditor.wpf.cs (WPF: AddOwner from TextView)
+		// or TextEditor.uno.cs (Uno: DependencyProperty.Register) for each platform.
 
 		/// <summary>
 		/// Gets/Sets the options currently used by the text editor.
@@ -185,34 +152,9 @@ namespace ICSharpCode.AvalonEdit
 
 		void OnOptionsChanged(TextEditorOptions oldValue, TextEditorOptions newValue)
 		{
-			if (oldValue != null) {
-				PropertyChangedWeakEventManager.RemoveListener(oldValue, this);
-			}
+			OnOptionsChangedCore(oldValue, newValue);
 			textArea.Options = newValue;
-			if (newValue != null) {
-				PropertyChangedWeakEventManager.AddListener(newValue, this);
-			}
 			OnOptionChanged(new PropertyChangedEventArgs(null));
-		}
-
-		/// <inheritdoc cref="IWeakEventListener.ReceiveWeakEvent"/>
-		protected virtual bool ReceiveWeakEvent(Type managerType, object sender, EventArgs e)
-		{
-			if (managerType == typeof(PropertyChangedWeakEventManager)) {
-				OnOptionChanged((PropertyChangedEventArgs)e);
-				return true;
-			} else if (managerType == typeof(TextDocumentWeakEventManager.TextChanged)) {
-				OnTextChanged(e);
-				return true;
-			} else if (managerType == typeof(PropertyChangedEventManager)) {
-				return HandleIsOriginalChanged((PropertyChangedEventArgs)e);
-			}
-			return false;
-		}
-
-		bool IWeakEventListener.ReceiveWeakEvent(Type managerType, object sender, EventArgs e)
-		{
-			return ReceiveWeakEvent(managerType, sender, e);
 		}
 		#endregion
 
@@ -220,7 +162,7 @@ namespace ICSharpCode.AvalonEdit
 		/// <summary>
 		/// Gets/Sets the text of the current document.
 		/// </summary>
-		[Localizability(LocalizationCategory.Text), DefaultValue("")]
+		[DefaultValue("")]
 		public string Text {
 			get {
 				TextDocument document = this.Document;
@@ -260,44 +202,19 @@ namespace ICSharpCode.AvalonEdit
 		}
 		#endregion
 
-		#region TextArea / ScrollViewer properties
+		#region TextArea property
 		readonly TextArea textArea;
-		ScrollViewer scrollViewer;
-
-		/// <summary>
-		/// Is called after the template was applied.
-		/// </summary>
-		public override void OnApplyTemplate()
-		{
-			base.OnApplyTemplate();
-			scrollViewer = (ScrollViewer)Template.FindName("PART_ScrollViewer", this);
-		}
 
 		/// <summary>
 		/// Gets the text area.
 		/// </summary>
 		public TextArea TextArea {
-			get {
-				return textArea;
-			}
+			get { return textArea; }
 		}
 
-		/// <summary>
-		/// Gets the scroll viewer used by the text editor.
-		/// This property can return null if the template has not been applied / does not contain a scroll viewer.
-		/// </summary>
-		internal ScrollViewer ScrollViewer {
-			get { return scrollViewer; }
-		}
-
-		bool CanExecute(RoutedUICommand command)
+		object IServiceProvider.GetService(Type serviceType)
 		{
-			return command.CanExecute(null, textArea);
-		}
-
-		void Execute(RoutedUICommand command)
-		{
-			command.Execute(null, textArea);
+			return textArea.GetService(serviceType);
 		}
 		#endregion
 
@@ -308,7 +225,6 @@ namespace ICSharpCode.AvalonEdit
 		public static readonly DependencyProperty SyntaxHighlightingProperty =
 			DependencyProperty.Register("SyntaxHighlighting", typeof(IHighlightingDefinition), typeof(TextEditor),
 										new FrameworkPropertyMetadata(OnSyntaxHighlightingChanged));
-
 
 		/// <summary>
 		/// Gets/sets the syntax highlighting definition used to colorize the text.
@@ -342,7 +258,6 @@ namespace ICSharpCode.AvalonEdit
 		/// Creates the highlighting colorizer for the specified highlighting definition.
 		/// Allows derived classes to provide custom colorizer implementations for special highlighting definitions.
 		/// </summary>
-		/// <returns></returns>
 		protected virtual IVisualLineTransformer CreateColorizer(IHighlightingDefinition highlightingDefinition)
 		{
 			if (highlightingDefinition == null)
@@ -398,11 +313,7 @@ namespace ICSharpCode.AvalonEdit
 					editor.TextArea.ReadOnlySectionProvider = ReadOnlySectionDocument.Instance;
 				else
 					editor.TextArea.ReadOnlySectionProvider = NoReadOnlySections.Instance;
-
-				TextEditorAutomationPeer peer = TextEditorAutomationPeer.FromElement(editor) as TextEditorAutomationPeer;
-				if (peer != null) {
-					peer.RaiseIsReadOnlyChanged((bool)e.OldValue, (bool)e.NewValue);
-				}
+				editor.OnIsReadOnlyChangedCore(editor, (bool)e.OldValue, (bool)e.NewValue);
 			}
 		}
 		#endregion
@@ -473,36 +384,13 @@ namespace ICSharpCode.AvalonEdit
 		static void OnShowLineNumbersChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
 		{
 			TextEditor editor = (TextEditor)d;
-			var leftMargins = editor.TextArea.LeftMargins;
-			if ((bool)e.NewValue) {
-				LineNumberMargin lineNumbers = new LineNumberMargin();
-				Line line = (Line)DottedLineMargin.Create();
-				leftMargins.Insert(0, lineNumbers);
-				leftMargins.Insert(1, line);
-				var lineNumbersForeground = new Binding("LineNumbersForeground") { Source = editor };
-				line.SetBinding(Line.StrokeProperty, lineNumbersForeground);
-				lineNumbers.SetBinding(Control.ForegroundProperty, lineNumbersForeground);
-			} else {
-				for (int i = 0; i < leftMargins.Count; i++) {
-					if (leftMargins[i] is LineNumberMargin) {
-						leftMargins.RemoveAt(i);
-						if (i < leftMargins.Count && DottedLineMargin.IsDottedLineMargin(leftMargins[i])) {
-							leftMargins.RemoveAt(i);
-						}
-						break;
-					}
-				}
-			}
+			editor.OnShowLineNumbersChangedCore(editor, (bool)e.OldValue, (bool)e.NewValue);
 		}
 		#endregion
 
 		#region LineNumbersForeground
-		/// <summary>
-		/// LineNumbersForeground dependency property.
-		/// </summary>
-		public static readonly DependencyProperty LineNumbersForegroundProperty =
-			DependencyProperty.Register("LineNumbersForeground", typeof(Brush), typeof(TextEditor),
-										new FrameworkPropertyMetadata(Brushes.Gray, OnLineNumbersForegroundChanged));
+		// LineNumbersForegroundProperty is declared in TextEditor.wpf.cs (default=Brushes.Gray, WPF-specific)
+		// or TextEditor.uno.cs for the Uno platform.
 
 		/// <summary>
 		/// Gets/sets the Brush used for displaying the foreground color of line numbers.
@@ -514,12 +402,7 @@ namespace ICSharpCode.AvalonEdit
 
 		static void OnLineNumbersForegroundChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
 		{
-			TextEditor editor = (TextEditor)d;
-			var lineNumberMargin = editor.TextArea.LeftMargins.FirstOrDefault(margin => margin is LineNumberMargin) as LineNumberMargin; ;
-
-			if (lineNumberMargin != null) {
-				lineNumberMargin.SetValue(Control.ForegroundProperty, e.NewValue);
-			}
+			((TextEditor)d).OnLineNumbersForegroundChangedCore((TextEditor)d, e.NewValue);
 		}
 		#endregion
 
@@ -542,36 +425,12 @@ namespace ICSharpCode.AvalonEdit
 		}
 
 		/// <summary>
-		/// Copies the current selection to the clipboard.
-		/// </summary>
-		public void Copy()
-		{
-			Execute(ApplicationCommands.Copy);
-		}
-
-		/// <summary>
-		/// Removes the current selection and copies it to the clipboard.
-		/// </summary>
-		public void Cut()
-		{
-			Execute(ApplicationCommands.Cut);
-		}
-
-		/// <summary>
 		/// Begins a group of document changes and returns an object that ends the group of document
 		/// changes when it is disposed.
 		/// </summary>
 		public IDisposable DeclareChangeBlock()
 		{
 			return GetDocument().RunUpdate();
-		}
-
-		/// <summary>
-		/// Removes the current selection without copying it to the clipboard.
-		/// </summary>
-		public void Delete()
-		{
-			Execute(ApplicationCommands.Delete);
 		}
 
 		/// <summary>
@@ -583,225 +442,11 @@ namespace ICSharpCode.AvalonEdit
 		}
 
 		/// <summary>
-		/// Scrolls one line down.
+		/// Clears the text.
 		/// </summary>
-		public void LineDown()
+		public void Clear()
 		{
-			if (scrollViewer != null)
-				scrollViewer.LineDown();
-		}
-
-		/// <summary>
-		/// Scrolls to the left.
-		/// </summary>
-		public void LineLeft()
-		{
-			if (scrollViewer != null)
-				scrollViewer.LineLeft();
-		}
-
-		/// <summary>
-		/// Scrolls to the right.
-		/// </summary>
-		public void LineRight()
-		{
-			if (scrollViewer != null)
-				scrollViewer.LineRight();
-		}
-
-		/// <summary>
-		/// Scrolls one line up.
-		/// </summary>
-		public void LineUp()
-		{
-			if (scrollViewer != null)
-				scrollViewer.LineUp();
-		}
-
-		/// <summary>
-		/// Scrolls one page down.
-		/// </summary>
-		public void PageDown()
-		{
-			if (scrollViewer != null)
-				scrollViewer.PageDown();
-		}
-
-		/// <summary>
-		/// Scrolls one page up.
-		/// </summary>
-		public void PageUp()
-		{
-			if (scrollViewer != null)
-				scrollViewer.PageUp();
-		}
-
-		/// <summary>
-		/// Scrolls one page left.
-		/// </summary>
-		public void PageLeft()
-		{
-			if (scrollViewer != null)
-				scrollViewer.PageLeft();
-		}
-
-		/// <summary>
-		/// Scrolls one page right.
-		/// </summary>
-		public void PageRight()
-		{
-			if (scrollViewer != null)
-				scrollViewer.PageRight();
-		}
-
-		/// <summary>
-		/// Pastes the clipboard content.
-		/// </summary>
-		public void Paste()
-		{
-			Execute(ApplicationCommands.Paste);
-		}
-
-		/// <summary>
-		/// Redoes the most recent undone command.
-		/// </summary>
-		/// <returns>True is the redo operation was successful, false is the redo stack is empty.</returns>
-		public bool Redo()
-		{
-			if (CanExecute(ApplicationCommands.Redo)) {
-				Execute(ApplicationCommands.Redo);
-				return true;
-			}
-			return false;
-		}
-
-		/// <summary>
-		/// Scrolls to the end of the document.
-		/// </summary>
-		public void ScrollToEnd()
-		{
-			ApplyTemplate(); // ensure scrollViewer is created
-			if (scrollViewer != null)
-				scrollViewer.ScrollToEnd();
-		}
-
-		/// <summary>
-		/// Scrolls to the start of the document.
-		/// </summary>
-		public void ScrollToHome()
-		{
-			ApplyTemplate(); // ensure scrollViewer is created
-			if (scrollViewer != null)
-				scrollViewer.ScrollToHome();
-		}
-
-		/// <summary>
-		/// Scrolls to the specified position in the document.
-		/// </summary>
-		public void ScrollToHorizontalOffset(double offset)
-		{
-			ApplyTemplate(); // ensure scrollViewer is created
-			if (scrollViewer != null)
-				scrollViewer.ScrollToHorizontalOffset(offset);
-		}
-
-		/// <summary>
-		/// Scrolls to the specified position in the document.
-		/// </summary>
-		public void ScrollToVerticalOffset(double offset)
-		{
-			ApplyTemplate(); // ensure scrollViewer is created
-			if (scrollViewer != null)
-				scrollViewer.ScrollToVerticalOffset(offset);
-		}
-
-		/// <summary>
-		/// Selects the entire text.
-		/// </summary>
-		public void SelectAll()
-		{
-			Execute(ApplicationCommands.SelectAll);
-		}
-
-		/// <summary>
-		/// Undoes the most recent command.
-		/// </summary>
-		/// <returns>True is the undo operation was successful, false is the undo stack is empty.</returns>
-		public bool Undo()
-		{
-			if (CanExecute(ApplicationCommands.Undo)) {
-				Execute(ApplicationCommands.Undo);
-				return true;
-			}
-			return false;
-		}
-
-		/// <summary>
-		/// Gets if the most recent undone command can be redone.
-		/// </summary>
-		public bool CanRedo {
-			get { return CanExecute(ApplicationCommands.Redo); }
-		}
-
-		/// <summary>
-		/// Gets if the most recent command can be undone.
-		/// </summary>
-		public bool CanUndo {
-			get { return CanExecute(ApplicationCommands.Undo); }
-		}
-
-		/// <summary>
-		/// Gets the vertical size of the document.
-		/// </summary>
-		public double ExtentHeight {
-			get {
-				return scrollViewer != null ? scrollViewer.ExtentHeight : 0;
-			}
-		}
-
-		/// <summary>
-		/// Gets the horizontal size of the current document region.
-		/// </summary>
-		public double ExtentWidth {
-			get {
-				return scrollViewer != null ? scrollViewer.ExtentWidth : 0;
-			}
-		}
-
-		/// <summary>
-		/// Gets the horizontal size of the viewport.
-		/// </summary>
-		public double ViewportHeight {
-			get {
-				return scrollViewer != null ? scrollViewer.ViewportHeight : 0;
-			}
-		}
-
-		/// <summary>
-		/// Gets the horizontal size of the viewport.
-		/// </summary>
-		public double ViewportWidth {
-			get {
-				return scrollViewer != null ? scrollViewer.ViewportWidth : 0;
-			}
-		}
-
-		/// <summary>
-		/// Gets the vertical scroll position.
-		/// </summary>
-		public double VerticalOffset {
-			get {
-				return scrollViewer != null ? scrollViewer.VerticalOffset : 0;
-			}
-		}
-
-		/// <summary>
-		/// Gets the horizontal scroll position.
-		/// </summary>
-		public double HorizontalOffset {
-			get {
-				return scrollViewer != null ? scrollViewer.HorizontalOffset : 0;
-			}
+			this.Text = string.Empty;
 		}
 		#endregion
 
@@ -837,12 +482,8 @@ namespace ICSharpCode.AvalonEdit
 		/// </summary>
 		[DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
 		public int CaretOffset {
-			get {
-				return textArea.Caret.Offset;
-			}
-			set {
-				textArea.Caret.Offset = value;
-			}
+			get { return textArea.Caret.Offset; }
+			set { textArea.Caret.Offset = value; }
 		}
 
 		/// <summary>
@@ -856,9 +497,7 @@ namespace ICSharpCode.AvalonEdit
 				else
 					return textArea.Selection.SurroundingSegment.Offset;
 			}
-			set {
-				Select(value, SelectionLength);
-			}
+			set { Select(value, SelectionLength); }
 		}
 
 		/// <summary>
@@ -872,9 +511,7 @@ namespace ICSharpCode.AvalonEdit
 				else
 					return 0;
 			}
-			set {
-				Select(SelectionStart, value);
-			}
+			set { Select(SelectionStart, value); }
 		}
 
 		/// <summary>
@@ -903,14 +540,6 @@ namespace ICSharpCode.AvalonEdit
 				else
 					return 1;
 			}
-		}
-
-		/// <summary>
-		/// Clears the text.
-		/// </summary>
-		public void Clear()
-		{
-			this.Text = string.Empty;
 		}
 		#endregion
 
@@ -993,185 +622,5 @@ namespace ICSharpCode.AvalonEdit
 			}
 		}
 		#endregion
-
-		#region MouseHover events
-		/// <summary>
-		/// The PreviewMouseHover event.
-		/// </summary>
-		public static readonly RoutedEvent PreviewMouseHoverEvent =
-			TextView.PreviewMouseHoverEvent.AddOwner(typeof(TextEditor));
-
-		/// <summary>
-		/// The MouseHover event.
-		/// </summary>
-		public static readonly RoutedEvent MouseHoverEvent =
-			TextView.MouseHoverEvent.AddOwner(typeof(TextEditor));
-
-
-		/// <summary>
-		/// The PreviewMouseHoverStopped event.
-		/// </summary>
-		public static readonly RoutedEvent PreviewMouseHoverStoppedEvent =
-			TextView.PreviewMouseHoverStoppedEvent.AddOwner(typeof(TextEditor));
-
-		/// <summary>
-		/// The MouseHoverStopped event.
-		/// </summary>
-		public static readonly RoutedEvent MouseHoverStoppedEvent =
-			TextView.MouseHoverStoppedEvent.AddOwner(typeof(TextEditor));
-
-
-		/// <summary>
-		/// Occurs when the mouse has hovered over a fixed location for some time.
-		/// </summary>
-		public event MouseEventHandler PreviewMouseHover {
-			add { AddHandler(PreviewMouseHoverEvent, value); }
-			remove { RemoveHandler(PreviewMouseHoverEvent, value); }
-		}
-
-		/// <summary>
-		/// Occurs when the mouse has hovered over a fixed location for some time.
-		/// </summary>
-		public event MouseEventHandler MouseHover {
-			add { AddHandler(MouseHoverEvent, value); }
-			remove { RemoveHandler(MouseHoverEvent, value); }
-		}
-
-		/// <summary>
-		/// Occurs when the mouse had previously hovered but now started moving again.
-		/// </summary>
-		public event MouseEventHandler PreviewMouseHoverStopped {
-			add { AddHandler(PreviewMouseHoverStoppedEvent, value); }
-			remove { RemoveHandler(PreviewMouseHoverStoppedEvent, value); }
-		}
-
-		/// <summary>
-		/// Occurs when the mouse had previously hovered but now started moving again.
-		/// </summary>
-		public event MouseEventHandler MouseHoverStopped {
-			add { AddHandler(MouseHoverStoppedEvent, value); }
-			remove { RemoveHandler(MouseHoverStoppedEvent, value); }
-		}
-		#endregion
-
-		#region ScrollBarVisibility
-		/// <summary>
-		/// Dependency property for <see cref="HorizontalScrollBarVisibility"/>
-		/// </summary>
-		public static readonly DependencyProperty HorizontalScrollBarVisibilityProperty = ScrollViewer.HorizontalScrollBarVisibilityProperty.AddOwner(typeof(TextEditor), new FrameworkPropertyMetadata(ScrollBarVisibility.Visible));
-
-		/// <summary>
-		/// Gets/Sets the horizontal scroll bar visibility.
-		/// </summary>
-		public ScrollBarVisibility HorizontalScrollBarVisibility {
-			get { return (ScrollBarVisibility)GetValue(HorizontalScrollBarVisibilityProperty); }
-			set { SetValue(HorizontalScrollBarVisibilityProperty, value); }
-		}
-
-		/// <summary>
-		/// Dependency property for <see cref="VerticalScrollBarVisibility"/>
-		/// </summary>
-		public static readonly DependencyProperty VerticalScrollBarVisibilityProperty = ScrollViewer.VerticalScrollBarVisibilityProperty.AddOwner(typeof(TextEditor), new FrameworkPropertyMetadata(ScrollBarVisibility.Visible));
-
-		/// <summary>
-		/// Gets/Sets the vertical scroll bar visibility.
-		/// </summary>
-		public ScrollBarVisibility VerticalScrollBarVisibility {
-			get { return (ScrollBarVisibility)GetValue(VerticalScrollBarVisibilityProperty); }
-			set { SetValue(VerticalScrollBarVisibilityProperty, value); }
-		}
-		#endregion
-
-		object IServiceProvider.GetService(Type serviceType)
-		{
-			return textArea.GetService(serviceType);
-		}
-
-		/// <summary>
-		/// Gets the text view position from a point inside the editor.
-		/// </summary>
-		/// <param name="point">The position, relative to top left
-		/// corner of TextEditor control</param>
-		/// <returns>The text view position, or null if the point is outside the document.</returns>
-		public TextViewPosition? GetPositionFromPoint(Point point)
-		{
-			if (this.Document == null)
-				return null;
-			TextView textView = this.TextArea.TextView;
-			return textView.GetPosition(TranslatePoint(point, textView) + textView.ScrollOffset);
-		}
-
-		/// <summary>
-		/// Scrolls to the specified line.
-		/// This method requires that the TextEditor was already assigned a size (WPF layout must have run prior).
-		/// </summary>
-		public void ScrollToLine(int line)
-		{
-			ScrollTo(line, -1);
-		}
-
-		/// <summary>
-		/// Scrolls to the specified line/column.
-		/// This method requires that the TextEditor was already assigned a size (WPF layout must have run prior).
-		/// </summary>
-		public void ScrollTo(int line, int column)
-		{
-			const double MinimumScrollFraction = 0.3;
-			ScrollTo(line, column, VisualYPosition.LineMiddle, null != scrollViewer ? scrollViewer.ViewportHeight / 2 : 0.0, MinimumScrollFraction);
-		}
-
-		/// <summary>
-		/// Scrolls to the specified line/column.
-		/// This method requires that the TextEditor was already assigned a size (WPF layout must have run prior).
-		/// </summary>
-		/// <param name="line">Line to scroll to.</param>
-		/// <param name="column">Column to scroll to (important if wrapping is 'on', and for the horizontal scroll position).</param>
-		/// <param name="yPositionMode">The mode how to reference the Y position of the line.</param>
-		/// <param name="referencedVerticalViewPortOffset">Offset from the top of the viewport to where the referenced line/column should be positioned.</param>
-		/// <param name="minimumScrollFraction">The minimum vertical and/or horizontal scroll offset, expressed as fraction of the height or width of the viewport window, respectively.</param>
-		public void ScrollTo(int line, int column, VisualYPosition yPositionMode, double referencedVerticalViewPortOffset, double minimumScrollFraction)
-		{
-			TextView textView = textArea.TextView;
-			TextDocument document = textView.Document;
-			if (scrollViewer != null && document != null) {
-				if (line < 1)
-					line = 1;
-				if (line > document.LineCount)
-					line = document.LineCount;
-
-				IScrollInfo scrollInfo = textView;
-				if (!scrollInfo.CanHorizontallyScroll) {
-					// Word wrap is enabled. Ensure that we have up-to-date info about line height so that we scroll
-					// to the correct position.
-					// This avoids that the user has to repeat the ScrollTo() call several times when there are very long lines.
-					VisualLine vl = textView.GetOrConstructVisualLine(document.GetLineByNumber(line));
-					double remainingHeight = referencedVerticalViewPortOffset;
-
-					while (remainingHeight > 0) {
-						DocumentLine prevLine = vl.FirstDocumentLine.PreviousLine;
-						if (prevLine == null)
-							break;
-						vl = textView.GetOrConstructVisualLine(prevLine);
-						remainingHeight -= vl.Height;
-					}
-				}
-				
-				Point p = textArea.TextView.GetVisualPosition(new TextViewPosition(line, Math.Max(1, column)), yPositionMode);
-				double verticalPos = p.Y - referencedVerticalViewPortOffset;
-				if (Math.Abs(verticalPos - scrollViewer.VerticalOffset) > minimumScrollFraction * scrollViewer.ViewportHeight) {
-					scrollViewer.ScrollToVerticalOffset(Math.Max(0, verticalPos));
-				}
-				if (column > 0) {
-					if (p.X > scrollViewer.ViewportWidth - Caret.MinimumDistanceToViewBorder * 2) {
-						double horizontalPos = Math.Max(0, p.X - scrollViewer.ViewportWidth / 2);
-						if (Math.Abs(horizontalPos - scrollViewer.HorizontalOffset) > minimumScrollFraction * scrollViewer.ViewportWidth) {
-							scrollViewer.ScrollToHorizontalOffset(horizontalPos);
-						}
-					} else {
-						scrollViewer.ScrollToHorizontalOffset(0);
-					}
-				}
-			}
-		}
 	}
 }
